refactor: move reusable notification components
