+++ gravitee-apim-console-webui/src/components/notification/notification-add-dialog/notification-add-dialog.component.scss
