+++ gravitee-apim-console-webui/src/components/notification/notification-edit-dialog/notification-edit-dialog.component.scss
@@ -2,7 +2,7 @@
 
 @use '@angular/material' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
-@use '../../../../../scss/gio-layout' as gio-layout;
+@use '../../../scss/gio-layout' as gio-layout;
 
 $typography: map.get(gio.$mat-theme, typography);
